1. Re:[맘스터치] POS-SAP 직영점 매출 데이터 연동 논의 미팅 개최 요청의 件 	1. [맘스터치] - [정산] - [POS내역수신(매출)] 	2. [맘스터치] - [정산] - [POS내역수신(결제)] 		- 신규 화면 개발 	3. [맘스터치] - [상품매출분석] - [매출상세현황(채널별)] - [매출상세현황(채널별) 탭] 		- 원복
2. [네이버 주문연동] 네이버 주문 매장연동 및 메뉴 매핑
	1. [네이버플레이스] - [네이버플레이스] - [네이버 주문연동]
		- 업체 리스트 조회시, ceoName(대표자명) 제거
	2. [네이버플레이스] - [네이버플레이스] - [네이버 메뉴관리]
		- 상품, 상품코드 그리드 정렬 기능 제거
    - 메인상품 체크박스 노출 및 삭제 기능 추가
    - 메인상품 순서 변경 불가 처리, 항상 최상위 순서 유지하도록 수정

3. [네이버] 네이버 주문연동 초기화 기능 추가
	1. [시스템관리] - [연동] - [네이버플레이스현황]
	2. [네이버플레이스] - [네이버플레이스] - [네이버플레이스 플러스 연동]
	3. [네이버플레이스] - [네이버플레이스] - [네이버주문연동]
		- 네이버플레이스 초기화, 네이버주문연동 초기화 개발

+++ src/main/java/kr/co/solbipos/sys/link/naverPlaceStatus/service/impl/NaverPlaceStatusServiceImpl.java
@@ -67,13 +67,23 @@ public class NaverPlaceStatusServiceImpl implements NaverPlaceStatusService {
 
         int result = 0;
 
-        // 네.아.로 ID, 네이버플레이스 연동매장 삭제
-        result = naverPlaceStatusMapper.deleteNaverLink(naverPlaceStatusVO);
-        if(result < 0) throw new JsonException(Status.SERVER_ERROR, messageService.get("cmm.saveFail"));
+        if("naverPlace".equals(naverPlaceStatusVO.getInFg())){
 
-        // 네이버 동의항목 삭제
-        result = naverPlaceStatusMapper.deleteNaverAgreement(naverPlaceStatusVO);
-        if(result < 0) throw new JsonException(Status.SERVER_ERROR, messageService.get("cmm.saveFail"));
+            // 네이버플레이스 연동정보 초기화
+            result = naverPlaceStatusMapper.updateNaverPlaceLink(naverPlaceStatusVO);
+            if(result < 0) throw new JsonException(Status.SERVER_ERROR, messageService.get("cmm.saveFail"));
+
+            // 네이버 동의항목 삭제
+            result = naverPlaceStatusMapper.deleteNaverAgreement(naverPlaceStatusVO);
+            if(result < 0) throw new JsonException(Status.SERVER_ERROR, messageService.get("cmm.saveFail"));
+        };
+
+        if("naverOrder".equals(naverPlaceStatusVO.getInFg())){
+
+            // 네이버주문연동 연동정보 초기화
+            result = naverPlaceStatusMapper.updateNaverOrderLink(naverPlaceStatusVO);
+            if(result < 0) throw new JsonException(Status.SERVER_ERROR, messageService.get("cmm.saveFail"));
+        };
 
         return result;
     }
